fix: 변수명 수정

+++ spring-security/springsecuritysecion6/src/main/java/com/eazybytes/controller/LoansController.java
@@ -15,7 +15,7 @@ public class LoansController {
     private final LoanRepository loanRepository;
 
     @GetMapping("/myLoans")
-    public List<Loans> getLoanDetails(@RequestParam int id) {
+    public List<Loans> getLoanDetails(@RequestParam Long id) {
         return loanRepository.findByCustomerIdOrderByStartDtDesc(id);
     }
