+++ spring-security/springsecuritysecion6/src/main/java/com/eazybytes/repository/LoanRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LoanRepository extends JpaRepository<Loans, Long> {
 
-    List<Loans> findByCustomerIdOrderByStartDtDesc(int customerId);
+    List<Loans> findByCustomerIdOrderByStartDtDesc(Long customerId);
 }
